requirements.txt
Django>=4.2,<5.1
django-environ>=0.11.0
whitenoise[brotli]>=6.6.0
mysqlclient>=2.2.0; sys_platform != "win32"
# On Windows use: pip install mysqlclient (or use PyMySQL as alternative)
Pillow>=10.0.0
django-ckeditor>=6.7.0
